% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_stocking_zones.R
\name{load_stocking_zones}
\alias{load_stocking_zones}
\title{Load Stocking Zones shapefile}
\usage{
load_stocking_zones()
}
\value{
An sf object
}
\description{
Loads the stocking zones GeoPackage as an sf object.
}
